+++ torch/csrc/jit/codegen/cuda/kernel_cache.cpp
@@ -306,6 +306,21 @@ std::vector<at::Tensor> FusionExecutorCache::runFusionWithInputs(
     const at::ArrayRef<IValue>& inputs) {
   FUSER_PERF_SCOPE("runFusionWithInputs");
 
+  auto detect_normalization_fusion = [&]() {
+    for (auto expr : fusion_->unordered_exprs()) {
+      if (expr->getExprType() == ExprType::BroadcastOp) {
+        auto output = expr->output(0);
+        auto input_origin_expr = expr->input(0)->getOrigin();
+        if (!fusion_->unordered_uses(output).empty() &&
+            input_origin_expr != nullptr &&
+            input_origin_expr->getExprType() == ExprType::ReductionOp) {
+          return true;
+        }
+      }
+    }
+    return false;
+  };
+
   LaunchParams launch_params;
 
   // get unique id `unique_id` for given input set `inputs`;
@@ -324,16 +339,12 @@ std::vector<at::Tensor> FusionExecutorCache::runFusionWithInputs(
 
     // caching strategy is different for pw-fusion and reduction-fusion.
     if (has_reduction_) {
+      bool isNormalizationFusion = detect_normalization_fusion();
       // Generate the reduction parameters
-      /*
-      auto reduction_params = (reduction_tv_.size() > 1)
-          ? getMultipleReductionHeuristics(fusion_.get(), inputs, reduction_tv_)
+      auto reduction_params = (isNormalizationFusion)
+          ? getNormalizationHeuristics(fusion_.get(), inputs, reduction_tv_)
           : getReductionHeuristics(
                 fusion_.get(), inputs, reduction_tv_.front());
-      */
-
-      auto reduction_params =
-          getMultipleReductionHeuristics(fusion_.get(), inputs, reduction_tv_);
 
       TORCH_INTERNAL_ASSERT(
           reduction_params.has_value(),
@@ -353,15 +364,15 @@ std::vector<at::Tensor> FusionExecutorCache::runFusionWithInputs(
         Fusion fusion_clone = *fusion_;
         FusionGuard fg(&fusion_clone);
 
+        // Separate the reduction TensorViews from the other TensorViews
+        // Ignore input TensorViews
         // Heavy weight call
+        std::vector<TensorView*> clone_reduction_tv;
+        std::vector<TensorView*> clone_other_tv;
         auto all_values = DependencyCheck::getAllValsBetween(
             {fusion_clone.inputs().begin(), fusion_clone.inputs().end()},
             fusion_clone.outputs());
 
-        // Separate the reduction TensorViews from the other TensorViews
-        // Ignore input TensorViews
-        std::vector<TensorView*> clone_reduction_tv;
-        std::vector<TensorView*> clone_other_tv;
         for (auto tv : ir_utils::filterByType<TensorView>(all_values)) {
           if (tv->hasReduction()) {
             clone_reduction_tv.push_back(tv);
@@ -370,15 +381,8 @@ std::vector<at::Tensor> FusionExecutorCache::runFusionWithInputs(
           }
         }
 
-        scheduleMultipleReduction(
-            &fusion_clone,
-            reduction_params.value(),
-            clone_reduction_tv,
-            clone_other_tv);
-
-        /*
-        if (clone_reduction_tv.size() > 1) {
-          scheduleMultipleReduction(
+        if (isNormalizationFusion) {
+          scheduleNormalization(
               &fusion_clone,
               reduction_params.value(),
               clone_reduction_tv,
@@ -402,7 +406,6 @@ std::vector<at::Tensor> FusionExecutorCache::runFusionWithInputs(
               single_reduction_tv,
               tv_outputs_of_reduction);
         }
-        */
 
         // This means we have not found a previously generated kernel that is
         // compatible with the new reduction params. We need to finish codegen.
@@ -518,7 +521,6 @@ void GraphCache::createFusion(const std::shared_ptr<Graph>& graph) {
           permuted_vec_optional_stride,
           type->requires_grad());
     }; // closing lambda
-
     for (auto input : graph->inputs()) {
       if (auto input_type = input->type()->cast<TensorType>()) {
         input->setType(type_permute_fn(input_type));
