+++ torch/csrc/jit/codegen/cuda/shape_inference.cpp
@@ -188,6 +188,13 @@ class NaiveTypePropagator {
         node->output(2)->setType(out_type);
         break;
       }
+      case aten::native_layer_norm_backward: {
+        auto out_type = node->input(0)->type()->cast<TensorType>();
+        node->output(0)->setType(out_type);
+        node->output(1)->setType(out_type);
+        node->output(2)->setType(out_type);
+        break;
+      }
       case aten::softmax: {
         auto out_type = node->input(0)->type()->cast<TensorType>();
